ARROW-3698: [Gandiva] Segmentation fault when using a large table in Gandiva
https://issues.apache.org/jira/browse/ARROW-3698

Author: suquark <zsy9509@mail.ustc.edu.cn>

Closes #2902 from suquark/gandiva and squashes the following commits:

9c7ac4ac6 <suquark> typo
24904cb9f <suquark> lint
a1650aa7b <suquark> fix test
f76cb64e4 <suquark> fix types
72ef82197 <suquark> fix types
fdb777560 <suquark> lint
a30449fac <suquark> Add tests for selection vector.
cb4fc5de1 <suquark> Add test for huge filters.
1c3dd3562 <suquark> int64 selection vector support
89d2f5e9c <suquark> Add tests for huge table.
9f9928e7d <suquark> revert changes to ZSTD
0566a9cf0 <suquark> remote verbose
c51fb5c71 <suquark> fix include
7547de33d <suquark> Lint. Better error message.
44886728c <suquark> verbose
d005c2223 <suquark> fix typo
307411ec9 <suquark> retrigger ci
706ccaa29 <suquark> fix jni types
ca0338c99 <suquark> More fixes
4e4c8a574 <suquark> fix eval_batch.h & lint
207d79230 <suquark> Fix more possible integer overflows.
c4daa9945 <suquark> Fix integer overflow in Gandiva which causes segfault.

+++ cpp/src/gandiva/local_bitmaps_holder.h
@@ -32,11 +32,11 @@ namespace gandiva {
 /// expression evaluation.
 class LocalBitMapsHolder {
  public:
-  LocalBitMapsHolder(int num_records, int num_local_bitmaps);
+  LocalBitMapsHolder(int64_t num_records, int num_local_bitmaps);
 
   int GetNumLocalBitMaps() const { return static_cast<int>(local_bitmaps_vec_.size()); }
 
-  int GetLocalBitMapSize() const { return local_bitmap_size_; }
+  int64_t GetLocalBitMapSize() const { return local_bitmap_size_; }
 
   uint8_t** GetLocalBitMapArray() const { return local_bitmaps_array_.get(); }
 
@@ -47,7 +47,7 @@ class LocalBitMapsHolder {
 
  private:
   /// number of records in the current batch.
-  int num_records_;
+  int64_t num_records_;
 
   /// A container of 'local_bitmaps_', each sized to accomodate 'num_records'.
   std::vector<std::unique_ptr<uint8_t>> local_bitmaps_vec_;
@@ -55,10 +55,10 @@ class LocalBitMapsHolder {
   /// An array of the local bitmaps.
   std::unique_ptr<uint8_t*> local_bitmaps_array_;
 
-  int local_bitmap_size_;
+  int64_t local_bitmap_size_;
 };
 
-inline LocalBitMapsHolder::LocalBitMapsHolder(int num_records, int num_local_bitmaps)
+inline LocalBitMapsHolder::LocalBitMapsHolder(int64_t num_records, int num_local_bitmaps)
     : num_records_(num_records) {
   // alloc an array for the pointers to the bitmaps.
   if (num_local_bitmaps > 0) {
@@ -66,7 +66,7 @@ inline LocalBitMapsHolder::LocalBitMapsHolder(int num_records, int num_local_bit
   }
 
   // 64-bit aligned bitmaps.
-  uint32_t roundUp64Multiple = (num_records_ + 63) >> 6;
+  int64_t roundUp64Multiple = (num_records_ + 63) >> 6;
   local_bitmap_size_ = roundUp64Multiple * 8;
 
   // Alloc 'num_local_bitmaps_' number of bitmaps, each of capacity 'num_records_'.
