+++ cpp/src/gandiva/projector.h
@@ -82,12 +82,12 @@ class Projector {
             const FieldVector& output_fields, std::shared_ptr<Configuration>);
 
   /// Allocate an ArrowData of length 'length'.
-  Status AllocArrayData(const DataTypePtr& type, int length, arrow::MemoryPool* pool,
+  Status AllocArrayData(const DataTypePtr& type, int64_t length, arrow::MemoryPool* pool,
                         ArrayDataPtr* array_data);
 
   /// Validate that the ArrayData has sufficient capacity to accomodate 'num_records'.
   Status ValidateArrayDataCapacity(const arrow::ArrayData& array_data,
-                                   const arrow::Field& field, int num_records);
+                                   const arrow::Field& field, int64_t num_records);
 
   /// Validate the common args for Evaluate() APIs.
   Status ValidateEvaluateArgsCommon(const arrow::RecordBatch& batch);
